migrated all the models into a models namespace
removed the neutral and failure tracking as these are calcuable by the db

corrected triumph and despair calculations (they can be cancelled by success or failure)

split the calculator from the summary and moved some common features to the roll container object

renamed the dieresult object or roll result to be a little more consistent

removed the old and unused data models from the app

modified the program to calculate the positive dice separately from the negative dice and then brute force their combined outcome pools adding a db record for each.

removed the code to save the data to a csv.