renamed the component object to match the rest of the dice statistics
reworked the model state code to split the next search dice from the current dice

added some pool information data to contextualize the numbers

added a search layout section and frameworked a dice rendering system.  dice are sorted and rendered in an array of icon elements.

added css for a die stroke

updated the icon set so it would work properly.

started work on submitting dice to the api for processing.