added a new pool combination constructor taking both pool halves.
updated the alternate total sum as the frequency will not overflow

renamed the pool calculators to outcome____ as the name makes a little more sense.

created a chart utility for managing rendering of the graphs, more cleanup needed.

added more responsive css.

added the alternate average line to graphs with it's own scale, scales not appearing on the triumph distribution.